Remove smoothing filter on motor speeds
Previously we were doing a simple low-pass filter on motor speeds. This didn't really work because the filter rate was not tied to the timer correctly.
Fixed a performance bug in mean_l that was making the low-pass filter work correctly (funny, that).
Consolidated motor enums.

+++ Avatar_Micro_Robot/Robot_V2/Power_Board/src/stdfunctions.c
@@ -23,7 +23,8 @@ int mean(size_t count, int *values) {
 }
 
 long mean_l(size_t count, long *values) {
-    long long total = 0;
+    // Note: I would total into a long long instead, but performance was surprisingly bad on PIC24!
+    long total = 0;
     size_t i;
     for (i = 0; i < count; i++) {
         total += values[i];
